+++ src/TokenSystem.Tests/TestTokenManager.cs
@@ -78,7 +78,7 @@ namespace TokenSystem.Tests
 
 		[Theory]
 		[InlineData(100)]
-		public void Mint_WhenPassedValidAmountAndAddresses_MintsTokensCorrectly(BigInteger amount)
+		public void Mint_WhenPassedValidAmountAndAddresses_MintsTokensCorrectly(int amount)
 		{
 			Address receiver = this.addresses[0];
 
@@ -91,7 +91,7 @@ namespace TokenSystem.Tests
 		[Theory]
 		[InlineData(0)]
 		[InlineData(-100)]
-		public void Mint_WhenPassedNonPositiveAmount_Throws(BigInteger amount)
+		public void Mint_WhenPassedNonPositiveAmount_Throws(int amount)
 		{
 			Address receiver = this.addresses[0];
 			Assert.Throws<NonPositiveTokenAmountException>(() => this.MintTokens(amount, receiver));
@@ -99,8 +99,8 @@ namespace TokenSystem.Tests
 
 		[Theory]
 		[InlineData(1000, 50)]
-		public void Transfer_WhenPassedValidAmount_TransfersTokensCorrectly(BigInteger mintAmount,
-			BigInteger transferAmount)
+		public void Transfer_WhenPassedValidAmount_TransfersTokensCorrectly(int mintAmount,
+			int transferAmount)
 		{
 			Address from = this.addresses[0];
 			Address to = this.addresses[1];
@@ -125,7 +125,7 @@ namespace TokenSystem.Tests
 		[Theory]
 		[InlineData(0)]
 		[InlineData(-234)]
-		public void Transfer_WhenPassedNonPositiveAmount_Throws(BigInteger transferAmount)
+		public void Transfer_WhenPassedNonPositiveAmount_Throws(int transferAmount)
 		{
 			Address from = this.addresses[0];
 			Address to = this.addresses[1];
@@ -137,8 +137,8 @@ namespace TokenSystem.Tests
 		[Theory]
 		[InlineData(100, 5000)]
 		public void Transfer_WhenPassedMoreThanOwned_Throws(
-			BigInteger mintAmount,
-			BigInteger transferAmount)
+			int mintAmount,
+			int transferAmount)
 		{
 			Address from = this.addresses[0];
 			Address to = this.addresses[1];
@@ -153,8 +153,8 @@ namespace TokenSystem.Tests
 		public void Transfer_WhenSenderAttemptsToTransferToHimself_Throws()
 		{
 			Address from = this.addresses[0];
-			BigInteger mintAmount = 100;
-			BigInteger transferAmount = 50;
+			int mintAmount = 100;
+			int transferAmount = 50;
 
 			this.MintTokens(mintAmount, from);
 
@@ -164,7 +164,7 @@ namespace TokenSystem.Tests
 
 		[Theory]
 		[InlineData(100, 90)]
-		public void Burn_WhenPassedValidArguments_BurnsCorrectly(BigInteger mintAmount, BigInteger burnAmount)
+		public void Burn_WhenPassedValidArguments_BurnsCorrectly(int mintAmount, int burnAmount)
 		{
 			Address address = this.addresses[0];
 
@@ -181,7 +181,7 @@ namespace TokenSystem.Tests
 		[Theory]
 		[InlineData(0)]
 		[InlineData(-1000)]
-		public void Burn_WhenPassedNonPositiveAmounts_Throws(BigInteger burnAmount)
+		public void Burn_WhenPassedNonPositiveAmounts_Throws(int burnAmount)
 		{
 			Address address = this.addresses[0];
 			Assert.Throws<NonPositiveTokenAmountException>(
@@ -190,7 +190,7 @@ namespace TokenSystem.Tests
 
 		[Theory]
 		[InlineData(100, 110)]
-		public void Burn_WhenAttemptingToBurnMoreThanOwned_Throws(BigInteger mintAmount, BigInteger burnAmount)
+		public void Burn_WhenAttemptingToBurnMoreThanOwned_Throws(int mintAmount, int burnAmount)
 		{
 			Address address = this.addresses[0];
 
@@ -202,7 +202,7 @@ namespace TokenSystem.Tests
 
 		[Theory]
 		[InlineData(1000)]
-		public void Mint_RaisesEvent(BigInteger mintAmount)
+		public void Mint_RaisesEvent(int mintAmount)
 		{
 			Address to = this.addresses[0];
 			this.tokenManager.TokensMinted += delegate(object sender, TokensMintedEventArgs<string> args)
@@ -216,7 +216,7 @@ namespace TokenSystem.Tests
 
 		[Theory]
 		[InlineData(1000, 100)]
-		public void Transfer_RaisesEvent(BigInteger mintAmount, BigInteger transferAmount)
+		public void Transfer_RaisesEvent(int mintAmount, int transferAmount)
 		{
 			Address from = this.addresses[0];
 			Address to = this.addresses[1];
@@ -234,7 +234,7 @@ namespace TokenSystem.Tests
 
 		[Theory]
 		[InlineData(1000, 100)]
-		public void Burn_RaisesEvent(BigInteger mintAmount, BigInteger burnAmount)
+		public void Burn_RaisesEvent(int mintAmount, int burnAmount)
 		{
 			Address from = this.addresses[0];
 
@@ -281,7 +281,7 @@ namespace TokenSystem.Tests
 		}
 
 		private void BurnTokens(
-			BigInteger amount,
+			int amount,
 			Address from,
 			ITokenPicker<string> tokenPicker = null)
 		{
